Database Architecture and Procedure Implementation for Transactional Systems/DataPilot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "networkx>=3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "streamlit>=1.47.1",
]
